Changed load_masked_signaltrack to use a precomputed mask if provided. Returns now also the number of read alignments (blocks), can be used to allocate array of appropriate size.
Correlation command: turns out that computing the correlation for more than a few million values is not feasible due to overflows that are not dealt with in scipy. To alleviate the situation, added iterator that finds consecutive blocks in arrays of positional indices (e.g. coming from a masked signal track), so correlation is computed on averages over these blocks (e.g. blocks of sequence conservation)